% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_sf.R
\docType{data}
\name{crs_sf}
\alias{crs_sf}
\title{Dataset of area extent polygons with CRS information}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 7628 rows and 8 columns.}
\usage{
data(crs_sf)
}
\description{
Used by the \code{crs_suggest()} function to look up suitable
             coordinate reference systems for input spatial data
}
\details{
Dataset of area extent polygons that correspond to CRS codes
}
\keyword{datasets}
