% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsuggest-package.R
\docType{package}
\name{crsuggest}
\alias{crsuggest}
\title{Get information on suggested coordinate reference systems for spatial data}
\description{
Get information on suggested coordinate reference systems for spatial data
}
\author{
Kyle Walker
}
