% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_crs.R
\name{guess_crs}
\alias{guess_crs}
\title{Guess the CRS of a dataset that is missing CRS information}
\usage{
guess_crs(input_sf, target_location, units = NULL, n_return = 10)
}
\arguments{
\item{input_sf}{An input sf object in a projected coordinate system that is
missing CRS information. For example, you may have loaded in a shapefile without
a .prj file, or your input data has no CRS definition attached.}

\item{target_location}{A coordinate pair of form \code{c(longitude, latitude)}
or an address/location that you know is located within your
input sf object.  If the mapboxapi package is installed, you can
supply a location name (e.g. an address or a city) instead of a
coordinate pair.}

\item{units}{If known, the units of your projected coordinate system (e.g. \code{"m"} for
meters or \code{"us-ft"} for US feet).  This is not required but will make
the guesses more accurate.}

\item{n_return}{The number of possible CRS choices to return; defaults to 10. A higher
number than that may include CRS options that are unlikely to work with
your data. Use the returned \code{dist_km} column to judge whether the CRS
guess makes sense for your data.}
}
\value{
A tibble of CRS guesses for your data, sorted in ascending order of distance between
        your target location and the input sf object's centroid when in that CRS.
}
\description{
Guess the CRS of a dataset that is missing CRS information
}
