% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_crs.R
\name{suggest_crs}
\alias{suggest_crs}
\title{Suggest coordinate systems for an input spatial dataset}
\usage{
suggest_crs(input, type = "projected", limit = 10, gcs = NULL, units = NULL)
}
\arguments{
\item{input}{A spatial dataset of class \code{"sf"}, \code{"Spatial*"}, or
\code{"RasterLayer"}.}

\item{type}{The output CRS type; defaults to \code{"projected"}.}

\item{limit}{How many results to return; defaults to \code{10}.}

\item{gcs}{(optional) The EPSG code for the corresponding geographic coordinate system of the results (e.g. \code{4326} for WGS 1984).}

\item{units}{(optional) The measurement units of the coordinate systems in the returned results.  Can be one of \code{"m"}, \code{"ft"}, or \code{"ft-us"}.}
}
\value{
A data frame with information about coordinate reference systems that could be suitably used for CRS transformation.
}
\description{
Suggest coordinate systems for an input spatial dataset
}
