% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsuggest-package.R
\docType{package}
\name{crsuggest}
\alias{crsuggest}
\title{Get information on suggested coordinate reference systems for spatial data}
\description{
Uses data from the EPSG Registry to look up suitable coordinate reference
system transformations for spatial datasets in R.  Returns a data frame with CRS codes
that can be used for CRS transformation and mapping projects.  Please see
the EPSG Dataset Terms of Use at \url{https://epsg.org/terms-of-use.html} for more information.
}
\author{
Kyle Walker
}
